requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.46
asyncpg==0.30.0
redis==5.2.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0